% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course.R
\name{compute_bearing_terminal}
\alias{compute_bearing_terminal}
\title{compute terminal bearing along a great circle path}
\usage{
compute_bearing_terminal(latitude1, longitude1, latitude2, longitude2)
}
\arguments{
\item{latitude1}{numeric vector of latitudes [degrees]}

\item{longitude1}{numeric vector of longitudes [degrees] corresponding with latitude1}

\item{latitude2}{numeric vector of latitudes [degrees]}

\item{longitude2}{numeric vector of longitudes [degrees] corresponding with latitude2}
}
\value{
numeric, bearing in degrees
}
\description{
This function provides a convienant wrapper to 
\code{\link[geosphere]{finalBearing}}
}

