% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{compute_dist_cross_track}
\alias{compute_dist_cross_track}
\title{compute spherical cross track distance (error)}
\usage{
compute_dist_cross_track(latitude, longitude, path_latitude1, path_longitude1,
  path_latitude2, path_longitude2)
}
\arguments{
\item{latitude}{numeric vector of latitudes [degrees]}

\item{longitude}{numeric vector of longitudes [degrees] corresponding with latitude}

\item{path_latitude1}{numeric vector of latitudes [degrees]}

\item{path_longitude1}{numeric vector of longitudes [degrees] corresponding with path_latitude1}

\item{path_latitude2}{numeric vector of latitudes [degrees]}

\item{path_longitude2}{numeric vector of longitudes [degrees] corresponding with path_latitude2}
}
\value{
numeric, cross track distance nautical miles
}
\description{
This function provides a convienant wrapper to 
\code{\link[geosphere]{dist2gc}}. Per that documentation, this measures
the distance from the input point(s) defined by \code{latitude} and 
\code{longitude} to the nearest location on the great circle
path defined by the \code{path_*} inputs.
}

