% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{compute_dist_rhumb}
\alias{compute_dist_rhumb}
\title{compute spherical distance along a rhumb line}
\usage{
compute_dist_rhumb(latitude1, longitude1, latitude2, longitude2)
}
\arguments{
\item{latitude1}{numeric vector of latitudes [degrees]}

\item{longitude1}{numeric vector of longitudes [degrees] corresponding with latitude1}

\item{latitude2}{numeric vector of latitudes [degrees]}

\item{longitude2}{numeric vector of longitudes [degrees] corresponding with latitude2}
}
\value{
numeric, distance in nautical miles
}
\description{
This function provides a convienant wrapper to 
\code{\link[geosphere]{distRhumb}}
}
