% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{compute_dist_to_polygon}
\alias{compute_dist_to_polygon}
\title{compute spherical distance to a line/polygon}
\usage{
compute_dist_to_polygon(latitude, longitude, poly_latitude, poly_longitude)
}
\arguments{
\item{latitude1}{numeric vector of latitudes [degrees]}

\item{longitude1}{numeric vector of longitudes [degrees] corresponding with latitude1}

\item{latitude2}{numeric vector of latitudes [degrees] defining a polygon (or line)}

\item{longitude2}{numeric vector of longitudes [degrees] corresponding with latitude2}

\item{latitude3}{numeric vector of latitudes [degrees]}

\item{longitude3}{numeric vector of longitudes [degrees] corresponding with latitude3}
}
\value{
numeric, distance along path in nautical miles
}
\description{
This function provides a convienant wrapper to 
\code{\link[geosphere]{dist2Line}}. Per that documentation, this measures
the distance from
}

