% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R, R/zzzz_et_names.R
\name{compute_get_earth_radius}
\alias{compute_get_earth_radius}
\alias{compute_set_earth_radius}
\alias{et_get_earth_radius}
\alias{et_set_earth_radius}
\title{get/set Earth's radius in nautical miles

Obviously the earth is not really a sphere, the
radius given here is an approximation such that
spherical geometry calculations return close to
accurate results}
\usage{
compute_get_earth_radius()

compute_set_earth_radius(radius)

et_get_earth_radius()

et_set_earth_radius(radius)
}
\arguments{
\item{radius}{Numeric scalar. The radius of the earth 
(or any other sphere you care about) in nautical miles.}
}
\value{
numeric, Earth's radius in nautical miles
}
\description{
get/set Earth's radius in nautical miles

Obviously the earth is not really a sphere, the
radius given here is an approximation such that
spherical geometry calculations return close to
accurate results
}
