% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_parsers.R, R/zzzz_et_names.R
\name{compute_parse_lldms}
\alias{compute_parse_lldms}
\alias{et_parse_lldms}
\title{parse coordinates from latitude/longitude minute-second-direction strings}
\usage{
compute_parse_lldms(strIn)

et_parse_lldms(strIn)
}
\description{
parse lat/lon strings of the form XXX-YY-ZZ.ZZZ<Direction>
}
\author{
Kevin Sprong <ksprong@mitre.org>
}
