% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R, R/zzzz_et_names.R
\name{compute_projection_rhumb}
\alias{compute_projection_rhumb}
\alias{et_projection_rhumb}
\title{compute location of coordinates after rhumb line projection}
\usage{
compute_projection_rhumb(latitude, longitude, bearing, distance,
  output_type = "data.table")

et_projection_rhumb(latitude, longitude, bearing, distance,
  output_type = "data.table")
}
\arguments{
\item{latitude}{numeric vector of latitudes [degrees]}

\item{longitude}{numeric vector of longitudes [degrees] corresponding with latitude}

\item{bearing}{initial bearing [degrees]}

\item{distance}{projection distance [nautical miles]}

\item{output_type}{string in \code{c("matrix", "data.table", "data.frame", "list")}}
}
\value{
a \code{data.table} with columns latitude and longitude. If the input coordinates have length 1, then a named numeric vector is returned if
}
\description{
**DEPRECATED** Please use \code{compute_projection(method="rhumb")}

This function provides a convienant wrapper to 
\code{\link[geosphere]{destPointRhumb}}
}
