% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geohash.r
\name{decode}
\alias{dcd}
\alias{decode}
\title{Decode a geohash to lat/lon}
\usage{
decode(x)

dcd(x)
}
\arguments{
\item{x}{the geohash string}
}
\description{
This function takes in a geohash string and returns a 3x2 matrix where the first 
column contains the decoded latitude information and the second contains the 
deocded longitude information. The first row contains the minimum value; the 
second row contains the maximum value and the third row contains the error. To ge the 
center of the geohash box, use the hash_center(x) function.

A faster version of decode
}
\author{
Seth Wenchel <swenchel@mitre.org>

Seth Wenchel <swenchel@mitre.org>
}

