% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geohash.r
\name{encode}
\alias{encode}
\title{encode point into a geohash string}
\usage{
encode(latitude, longitude, precision = 5)
}
\arguments{
\item{latitude}{numeric value of latitude}

\item{longitude}{numeric value of latitude}

\item{precision}{integer length of return value}
}
\value{
a string representation of the geohash of the point of length precision.
}
\description{
encode point into a geohash string
}
\examples{
# should produce "ezs42"
library("earthtools")
encode(42.6, -5.6)

}

