% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{et_get_earth_radius}
\alias{et_get_earth_radius}
\alias{et_set_earth_radius}
\title{get Earth's radius in nautical miles}
\usage{
et_get_earth_radius()

et_set_earth_radius(radius)
}
\arguments{
\item{radius}{Numeric scalar. The radius of the earth 
(or any other sphere you care about) in nautical miles.}
}
\value{
numeric, Earth's radius in nautical miles
}
\description{
Obviously the earth is not really a sphere, the
radius given here is an approximation such that
spherical geometry calculations return close to
accurate results
}
