% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_parsers.R
\name{et_parse_lls}
\alias{et_parse_lls}
\title{parse coordinates from latitude/longitude + direction strings}
\usage{
et_parse_lls(strIn)
}
\description{
parse lat or lon strings of the form nnnnnn.nnnn<Direction>
}
\author{
Kevin Sprong <ksprong@mitre.org>
}
