% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle.R
\name{et_turn_magnitude}
\alias{et_turn_magnitude}
\title{compute turn magnitude}
\usage{
et_turn_magnitude(angle1, angle2, turn_direction = "closest")
}
\arguments{
\item{angle1}{numeric vector, initial heading in degrees}

\item{angle2}{numeric vector, terminal heading in degrees}

\item{turn_direction}{character string, should be one of
\code{"closest", "left", "right"}}
}
\description{
Due to angles being measured in degrees modulo 360 it can be
annoying to determine the magnitude of a turn from, for example,
310 degrees to 30 degrees. This function simplifies the operation.
}
