% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_earth_radius}
\alias{get_earth_radius}
\alias{set_earth_radius}
\title{get/set Earth's radius in nautical miles

Obviously the earth is not really a sphere, the
radius given here is an approximation such that
spherical geometry calculations return close to
accurate results}
\usage{
get_earth_radius()

set_earth_radius(radius)
}
\arguments{
\item{radius}{Numeric scalar. The radius of the earth 
(or any other sphere you care about) in nautical miles.}
}
\value{
numeric, Earth's radius in nautical miles
}
\description{
get/set Earth's radius in nautical miles

Obviously the earth is not really a sphere, the
radius given here is an approximation such that
spherical geometry calculations return close to
accurate results
}
