% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geohash.r
\name{hash_center}
\alias{hash_center}
\alias{hc_cmp_mem}
\title{Returns the center of a geohash}
\usage{
hash_center(x)

hc_cmp_mem(x)
}
\arguments{
\item{x}{a string representing a geohash}
}
\description{
Given a geohash this function will calculate the center lat/lon coordinate.  
To get the bounds and the error, use the function dcd(x).

Use this version with repeated calls to hash_center when you 
might be decoding the same hash more than once.
}
\author{
Seth Wenchel <swenchel@mitre.org>

Seth Wenchel <swenchel@mitre.org>
}

