% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_parsers.R
\name{parse_lldms}
\alias{parse_lldms}
\title{parse coordinates from latitude/longitude minute-second-direction strings}
\usage{
parse_lldms(strIn)
}
\description{
parse lat/lon strings of the form XXX-YY-ZZ.ZZZ<Direction>
}
\author{
Kevin Sprong <ksprong@mitre.org>
}
