% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AviationUtils.R
\name{parse_lls}
\alias{parse_lls}
\title{parse coordinates from latitude/longitude + direction strings}
\usage{
parse_lls(strIn)
}
\description{
parse lat or lon strings of the form nnnnnn.nnnn<Direction>
}
\author{
Kevin Sprong <ksprong@mitre.org>
}

