% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{project_arc}
\alias{project_arc}
\title{produce a ring of pseudo-nodes}
\usage{
project_arc(center_latitude, center_longitude, radius, dist_between_points,
  tolerance = 0.01, from_angle = 0, to_angle = 359.999)
}
\arguments{
\item{center_latitude}{Numeric. The latitude of the range ring center point.}

\item{center_longitude}{Numeric. The longitude of the range ring center point.}

\item{radius}{Numeric. The radius of the arc about the center point in nautical miles.}

\item{dist_between_points}{Numeric. The maximum distance between points on the
range ring. Actual distance is between this upper bound and this minus the tolerance,
all in nautical miles.}

\item{tolerance}{Numeric. How close to the target distance between points do you need to get?}

\item{from_angle}{Numeric. The angle from the center point to the arc defining the left-most
point of the arc.}

\item{to_angle}{Numeric. The angle from the center point to the arc defining the right-most
point of the arc (moving clockwise from the \code{from_angle}).}
}
\description{
It can be useful to add a range ring of nodes to a graph. This function will select such
nodes subject to radius and disance constraints. As a general rule the max distance should
be twice the max fix distance tolerance so that aircraft "pass by" no more or less than one
fix on the ring (assuming they don't double back or fly in at a near tangent heading)
}
