% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_earth_radius}
\alias{get_earth_radius}
\title{get Earth's radius in nautical miles}
\usage{
get_earth_radius()
}
\value{
numeric, Earth's radius in nautical miles
}
\description{
Obviously the earth is not really a sphere, the
radius given here is an approximation such that
spherical geometry calculations return close to
accurate results
}

